test bit fix

+++ main.cpp
@@ -28,7 +28,8 @@ int main() {
 
     bitset<32> inputBits(binaryInput);
     //
-    unsigned int testText = static_cast<unsigned int>(inputBits.to_ulong());
+    //unsigned int testText = static_cast<unsigned int>(inputBits.to_ulong());
+    unsigned long testText = inputBits.to_ulong();
     bitset<48> outputBits = apply_expansion(testText);
 
     outputFile << "Initial Text: " << inputBits << endl;
